pyproject.toml
[tool.poetry]
name = "dir-diary"
version = "0.3.0"
description = "A Python CLI tool to summarize your software project folder with LLMs"
license = "MIT"
authors = ["Christopher Carroll Smith <christopher.smith@promptlytechnologies.com>"]
readme = "README.md"
packages = [{include = "dir_diary"}]

[tool.poetry.dependencies]
python = "^3.8.1"
langchain = "^0.0.272"
python-dotenv = "^1.0.0"
pydantic = "^2.3.0"
pathspec = "^0.11.2"
openai = "^0.28.0"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"

[tool.poetry.scripts]
summarize = "dir_diary.cli:cli"
summarise = "dir_diary.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
